requirements.txt
matplotlib==2.1.1
numpy==1.14.0
tensorflow==1.4.0
Pillow==5.0.0
